requirements.txt
flask>=3.0.0
flask-sqlalchemy>=3.1.1
flask-login>=0.6.3
flask-migrate>=4.0.5
requests>=2.31.0
python-dotenv>=1.0.0
flask-assets>=2.1.0
flask-limiter>=3.5.0
flask-caching>=2.1.0
flask-cors>=4.0.0
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
redis>=5.0.1
prometheus-flask-exporter>=0.23.0
flask-compress>=1.14
gunicorn>=21.2.0
